Create Equipment Component and apply React Router to the App
